refactor: startegy 패키지 분리
