+++ src/main/java/lotto/Application.java
@@ -1,7 +1,7 @@
 package lotto;
 
 import lotto.controller.LottoController;
-import lotto.domain.strategy.AutoBuy;
+import lotto.strategy.AutoBuy;
 
 public class Application {
     public static void main(String[] args) {
